+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 vNext
 -----
-* Breaking: Changed Numeric.Units.Dimensional.Prelude to export a dimensionally
-  typed `recip` instead of the one from `Prelude`.
+* Breaking: Changed Numeric.Units.Dimensional.Prelude to export dimensionally
+  typed `recip` and `logBase` instead of the ones from `Prelude`.
 * Breaking: Changed Numeric.Units.Dimensional.Prelude to export `(.)` and `id`
   from `Control.Category` instead of from `Prelude`.
 * Breaking: Changed the `HasDimension` typeclass to require an instance of the new
@@ -12,6 +12,10 @@ vNext
   module which may cause name collisions.
 * Breaking: Removed exports of `nMeter`, `nSecond`, `kilo`, etc from Numeric.Units.Dimensional.UnitNames.
   Access these instead by inspecting the relevant units or prefixes.
+* Breaking: Generalized the type of `dimensionlessLength` from
+  `(Num a, Foldable f) => f (Dimensional v d a) -> Dimensionless a)` to
+  `(Num a, Foldable f) => f b -> Dimensionless a`. This provides a weaker constraint on the type `a`
+  and may result in ambiguous types in code that depends on the former less general type.
 * Added `Data`, `Generic`, `Typeable` and `NFData` instances for many ancillary types.
 * Added `unQuantity` to the Coercion module to ease unwrapping without
   introducing ambiguous type variables.
@@ -19,6 +23,9 @@ vNext
 * Added a multiplicative `Monoid` instance for `AnyQuantity` and for `AnyUnit`.
 * Added the `DynQuantity` type to represent possibly valid quantities of statically
   unknown dimension, suitable for performing arithmetic with such quantities.
+* Added `nroot` function for term-level dimensions.
+* Added the Numeric.Units.Dimensional.Float module with convenient wrappers around functions
+  from RealFloat for inspecting floating point quantities.
 * Relocated git repository to https://github.com/bjornbm/dimensional
 
 1.0.1.1 (2015-11)